requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
gunicorn==21.2.0
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
